[OT-347] [REFACTOR]: profile api 리팩토링
